Update migrations, add tests, postman collection

+++ database/migrations/2022_08_10_105240_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedInteger('price');
+            $table->decimal('price');
             $table->boolean('published');
 
             $table->softDeletes();
